Rename utils module to formatters
